+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Version 0.2.3
 -------------
 * EntityManager offers two iterations now: Active only or All
 * EntityManager can pop entities, analogous to Dict.pop
+* Entities generated from Pixmap are dilated upon import by one pixel
 
 Version 0.2.2
 -------------
@@ -53,4 +54,3 @@ Misc
 Known Issues
 ^^^^^^^^^^^^
 * Occasionally the view can not be panned anymore using the middle mouse button in draw mode. Switching to select mode and back into draw mode reenabels panning again
-* Loading cell segments from an image directly is broken. As a workaround rotate the pixelmap by 90° counter clockwise and afterwards flip the image horizontially. Lets assume a transfromation function T(i) for the image i, then T(i) = flip(rot(i))